modified:   .gitignore 	new file:   examples/ddpg/car_racing/config.yaml 	new file:   examples/ddpg/car_racing/ddpg_car_racing.py 	new file:   requirements.txt 	new file:   rl_lib/src/algoritms/a2c/actor_critic.py 	modified:   rl_lib/src/algoritms/base_algo.py 	new file:   rl_lib/src/algoritms/ddpg/config.yaml 	modified:   rl_lib/src/algoritms/ddpg/ddpg.py 	modified:   rl_lib/src/algoritms/dqn/dqn.py 	modified:   rl_lib/src/algoritms/simple_q.py 	modified:   rl_lib/src/explore_env/exploration_manager.py 	new file:   rl_lib/src/explore_env/ou_noise.py 	modified:   rl_lib/src/explore_env/soft_q.py 	modified:   rl_lib/src/replay_buffers/random_buffers.py 	modified:   rl_lib/tests/config.yaml 	new file:   rl_lib/tests/ddpg_config.yaml 	renamed:    rl_lib/tests/drqn_config.yaml -> rl_lib/tests/dqn_config.yaml 	new file:   rl_lib/tests/first_test_ddpg.py 	modified:   rl_lib/tests/first_test_dqn.py 	modified:   rl_lib/tests/first_test_drqn.py 	modified:   setup.py

+++ rl_lib/src/algoritms/dqn/dqn.py
@@ -9,7 +9,7 @@ class DQN_Model(Model):
     super().__init__(model_config = config.get('model_config', {}), config = config,  default_config_path=__file__, **kwargs)
   
   def _prediction_processing(self, inputs: tf.Tensor, **kwargs):
-    mask = self.make_mask(kwargs['action'])
+    mask = self.make_mask(tf.cast(kwargs['action'], dtype = tf.int32))
     if len(inputs.shape) != len(mask.shape): mask = tf.expand_dims(mask, -1)
     return tf.reduce_sum(tf.multiply(inputs, mask), axis=kwargs['batch_dims'])
